requirements.txt
Pillow==8.1.2
requests==2.25.1
flake8==3.9.2
pytest==6.2.4
pytest-cov==2.12.0
pytest-mock==3.6.1
requests-mock==1.9.2

